implement hard concept
the disc destroy after 3 times bouncing the wall. The distroy is after 5 secounds (with this delay, farther the discus can hit ring until his destroing)
the discus destroy, when he not bounced 3 times the wall, but his velocity (speed) is 0
spawn discus after hit a ring and the player can immediately throw the discus
when all discus are distroy -> game over! (score and medal screen is appear)